refactor(user): rename folder
