+++ src/modules/auth/guard/authorization/roles.guard.ts
@@ -1,7 +1,7 @@
 import { CanActivate, ExecutionContext, Injectable } from "@nestjs/common";
 import { Reflector } from "@nestjs/core";
 import { ServiceException } from "../../../_common/filter/exception/service/serviceException";
-import { ROLES_KEY } from "../../../user/decorator/role";
+import { ROLES_KEY } from "../../../user/metadata/role";
 import { UserRole } from "../../../user/entity/user.entity";
 import { AUTH_GUARD_PAYLOAD } from "../const";
 import { Request } from "express";
