+++ src/modules/user/user.controller.ts
@@ -36,7 +36,7 @@ import {
 import { AUTH_GUARD_PAYLOAD } from "../auth/guard/const";
 import { DBQueryRunner } from "../db/query-runner/decorator/queryRunner.decorator";
 import { QueryRunnerInterceptor } from "../db/query-runner/queryRunner.interceptor";
-import { Roles } from "./decorator/role";
+import { Roles } from "./metadata/role";
 import { CreateUserDTO } from "./dto/request/createUserDTO";
 import { ReplacePassWordDTO } from "./dto/request/replacePwDTO";
 import { ReplaceRoleDTO } from "./dto/request/replaceRoleDTO";
